disable cgroups checks by default
The sole-member check is heuristical (e.g. we might be the first
member of a group intended for more processes) as well as limiting
(e.g. it might not matter that we're alone since the group may
only contain stuff where no clashes are guaranteed).

Perform a PID check by default and let people enable the cgroup
checks via param.